small corrections to code layout
